+++ opm/parser/eclipse/RawDeck/tests/StarTokenTests.cpp
@@ -106,10 +106,6 @@ BOOST_AUTO_TEST_CASE( readValueToken_basic_validity_tests ) {
     BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<double>("3.3d0"), 1e-6 );
     BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<double>("3.3E0"), 1e-6 );
     BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<double>("3.3D0"), 1e-6 );
-    BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<float>("3.3e0"), 1e-4 );
-    BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<float>("3.3d0"), 1e-4 );
-    BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<float>("3.3E0"), 1e-4 );
-    BOOST_CHECK_CLOSE( 3.3, Opm::readValueToken<float>("3.3D0"), 1e-4 );
     BOOST_CHECK_EQUAL( "OLGA", Opm::readValueToken<std::string>("OLGA") );
     BOOST_CHECK_EQUAL( "OLGA", Opm::readValueToken<std::string>("'OLGA'") );
     BOOST_CHECK_EQUAL( "123*456", Opm::readValueToken<std::string>("123*456") );
